requirements.txt
fastapi
pydantic
uvicorn
opentelemetry-exporter-jaeger
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
asyncio
aiohttp
requests